+++ src/app.module.ts
@@ -3,6 +3,9 @@ import { ConfigModule } from '@nestjs/config';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { TelegramService } from './service/telegram.service';
+import { CryptoService } from './service/crypto.service';
+import { MessageLogService } from './service/message-log.service';
+import { AiSummaryService } from './service/ai-summary.service';
 
 @Module({
   imports: [
@@ -12,6 +15,12 @@ import { TelegramService } from './service/telegram.service';
     }),
   ],
   controllers: [AppController],
-  providers: [AppService, TelegramService],
+  providers: [
+    AppService,
+    CryptoService,
+    MessageLogService,
+    AiSummaryService,
+    TelegramService,
+  ],
 })
 export class AppModule {}
